Original change: https://googleplex-android-review.googlesource.com/c/platform/frameworks/base/+/13827978

MUST ONLY BE SUBMITTED BY AUTOMERGER

+++ services/core/java/com/android/server/stats/pull/StatsPullAtomService.java
@@ -1486,7 +1486,7 @@ public class StatsPullAtomService extends SystemService {
         }
         for (int freqIndex = 0; freqIndex < timesMs.length; ++freqIndex) {
             int cluster = freqsClusters[freqIndex];
-            long freq = freqs[freqIndex];
+            int freq = (int) freqs[freqIndex];
             long timeMs = timesMs[freqIndex];
             pulledData.add(FrameworkStatsLog.buildStatsEvent(atomTag, cluster, freq, timeMs));
         }
